requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
itsdangerous
jinja2
psycopg2-binary
python-multipart
requests
passlib[bcrypt]
python-jose[cryptography]